requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
defopt==6.4.0
dnspython==2.4.1
docutils==0.20.1
email-validator==2.0.0.post2
executing==1.2.0
Faker==19.2.0
fastapi==0.100.0
filelock==3.12.2
fonttools==4.41.1
fsspec==2023.6.0
fvalues==0.0.4
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
Levenshtein==0.21.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
merge-args==0.1.5
more-itertools==10.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numerizer==0.2.3
numpy==1.25.1
orjson==3.9.2
ought-ice==0.5.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkg_resources==0.0.0
pockets==0.9.1
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.12
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.21.1
python-multipart==0.0.6
python-ulid==1.1.0
pytz==2023.3
PyYAML==6.0.1
questionary==1.10.0
rapidfuzz==3.1.2
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rouge-metric==1.0.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
sphinxcontrib-napoleon==0.7
stack-data==0.6.2
starlette==0.27.0
structlog==23.1.0
tenacity==8.2.2
thefuzz==0.19.0
threadpoolctl==3.2.0
tokenizers==0.13.3
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
